Split git snapshot cache tests
