+++ Packages/macOS/CmuxGit/Tests/CmuxGitTests/GitTrackedChangesSnapshotCacheTests.swift
@@ -1,0 +1,108 @@
+import Foundation
+import Testing
+@testable import CmuxGit
+
+@Suite struct GitTrackedChangesSnapshotCacheTests {
+    @Test func alternatingGenerationsKeepSeparateTrackedSnapshotCacheEntries() async throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        let entry = try fixture.writeWorkingTreeFile("file.txt", contents: "hello")
+        try fixture.writeIndex(GitIndexFixture(version: 2, entries: [entry]))
+        let repository = try #require(GitMetadataService.resolveGitRepository(containing: fixture.root.path))
+        let filePath = fixture.root.appendingPathComponent("file.txt").path
+        let reader = CountingGitFileStatusReader()
+        let service = GitMetadataService(fileStatusReader: reader)
+
+        _ = await service.gitTrackedChangesSnapshot(
+            repository: repository,
+            trackedPathEventGeneration: 40
+        )
+        _ = await service.gitTrackedChangesSnapshot(
+            repository: repository,
+            trackedPathEventGeneration: 41
+        )
+        _ = await service.gitTrackedChangesSnapshot(
+            repository: repository,
+            trackedPathEventGeneration: 40
+        )
+
+        #expect(reader.callCount(atPath: filePath) == 2)
+    }
+
+    @Test func refreshedCacheEntryMovesBehindOlderEntryForEviction() async throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        let repository = try #require(GitMetadataService.resolveGitRepository(containing: fixture.root.path))
+        let cache = GitTrackedChangesSnapshotCache(maximumEntryCount: 2)
+        let indexStatSignature = GitIndexStatSignature(
+            size: 1,
+            mtimeSeconds: 2,
+            mtimeNanoseconds: 3
+        )
+        let firstSnapshot = GitTrackedChangesSnapshot(
+            isDirty: false,
+            indexSignature: "first",
+            indexContentSignature: "first-content"
+        )
+        let refreshedFirstSnapshot = GitTrackedChangesSnapshot(
+            isDirty: true,
+            indexSignature: "first-refreshed",
+            indexContentSignature: "first-refreshed-content"
+        )
+        let secondSnapshot = GitTrackedChangesSnapshot(
+            isDirty: false,
+            indexSignature: "second",
+            indexContentSignature: "second-content"
+        )
+        let thirdSnapshot = GitTrackedChangesSnapshot(
+            isDirty: false,
+            indexSignature: "third",
+            indexContentSignature: "third-content"
+        )
+
+        await cache.store(
+            firstSnapshot,
+            repository: repository,
+            indexStatSignature: indexStatSignature,
+            trackedPathEventGeneration: 1
+        )
+        await cache.store(
+            secondSnapshot,
+            repository: repository,
+            indexStatSignature: indexStatSignature,
+            trackedPathEventGeneration: 2
+        )
+        await cache.store(
+            refreshedFirstSnapshot,
+            repository: repository,
+            indexStatSignature: indexStatSignature,
+            trackedPathEventGeneration: 1
+        )
+        await cache.store(
+            thirdSnapshot,
+            repository: repository,
+            indexStatSignature: indexStatSignature,
+            trackedPathEventGeneration: 3
+        )
+
+        let refreshedFirst = await cache.snapshot(
+            repository: repository,
+            indexStatSignature: indexStatSignature,
+            trackedPathEventGeneration: 1
+        )
+        let evictedSecond = await cache.snapshot(
+            repository: repository,
+            indexStatSignature: indexStatSignature,
+            trackedPathEventGeneration: 2
+        )
+        let third = await cache.snapshot(
+            repository: repository,
+            indexStatSignature: indexStatSignature,
+            trackedPathEventGeneration: 3
+        )
+
+        #expect(refreshedFirst == refreshedFirstSnapshot)
+        #expect(evictedSecond == nil)
+        #expect(third == thirdSnapshot)
+    }
+}
